renamed files split code over multiple headers replaced custArr by std::vector
